Update README and fix errors
