+++ README.md
@@ -1,1 +1,58 @@
-# enumerables
+# Enumerables
+
+This repository is dedicated to our project in which we created a custom enumerable list using Ruby 
+
+## Built With
+* Ruby
+
+## Getting Started
+
+Clone the repository to get a local copy
+
+### Prerequisites
+* Ruby
+
+## Setup
+
+Clone the repo locally on your machine by typing the following commands in your terminal/Command Prompt
+
+- `git clone https://github.com/yusufnviiri/enumerables.git`
+- `cd enumerables`
+- `ruby list.rb`
+
+## Authors
+
+👤 **Yusuf Nviiri**
+
+- GitHub: [@yusufnviiri](https://github.com/yusufnviiri)
+- Twitter: [@YNviiri](https://twitter.com/YNviiri)
+- LinkedIn: [yusufnviiri]( https://www.linkedin.com/in/yusuf-nviiri-8b4146206/)
+
+
+👤 **Talha Zuberi**
+
+- GitHub: [@tkz96](https://github.com/tkz96)
+- Twitter: [@TKZ96](https://twitter.com/tkz96)
+- LinkedIn: [LinkedIn](https://linkedin.com/in/talha-zuberi)
+
+Show your support
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](../../issues/).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Hat tip to anyone whose code was used
+- Inspiration
+- etc
+
+## 📝 License
+
+This project is [MIT](./MIT.md) licensed.
