+++ list.rb
@@ -1,20 +1,12 @@
-module MyEnumerable
-  def all?
-    condition = true
-    each { |x| condition = false unless yield(x) }
-    condition
+require_relative './myEnum'
+
+class MyList
+  include MyEnumerable
+  def initialize(*args)
+    @list = args
   end
 
-  def any?
-    condition = false
-    each { |x| condition = true if yield(x) }
-    condition
-  end
-
-  def filter?
-    list = []
-    each { |x| list << x if yield(x) }
-
-    list
+  def each(&block)
+    @list.each(&block)
   end
 end
